Update link to MSX2 Technical Handbook.
Konamiman originally published his transcript of MSX2 Tehnical Handbook
in the form of text files in his web site, but has later converted
it to Markdown files hosted in a dedicated GitHub repository.
This commit changes the link in msx.ts from the old text files
to the dedicated repository.